+++ parsed_json.go
@@ -7,8 +7,8 @@ import (
 	"io"
 	"io/ioutil"
 	"math"
-	"sync"
 	"strconv"
+	"sync"
 )
 
 //
@@ -44,7 +44,7 @@ type internalParsedJson struct {
 	ParsedJson
 	containing_scope_offset []uint64
 	isvalid                 bool
-	index_chan				chan indexChan
+	index_chan              chan indexChan
 }
 
 func (pj *internalParsedJson) initialize(size int) {
@@ -681,381 +681,6 @@ func (i *Iter) Array(dst *Array) (*Array, error) {
 	dst.off = i.off
 
 	return dst, nil
-}
-
-// Object represents a JSON object.
-type Object struct {
-	// Complete tape
-	tape ParsedJson
-
-	// offset of the next entry to be decoded
-	off int
-}
-
-// Map will unmarshal into a map[string]interface{}
-// See Iter.Interface() for a reference on value types.
-func (o *Object) Map(dst map[string]interface{}) (map[string]interface{}, error) {
-	if dst == nil {
-		dst = make(map[string]interface{})
-	}
-	var tmp Iter
-	for {
-		name, t, err := o.NextElement(&tmp)
-		if err != nil {
-			return nil, err
-		}
-		if t == TypeNone {
-			// Done
-			break
-		}
-		dst[name], err = tmp.Interface()
-		if err != nil {
-			return nil, fmt.Errorf("parsing element %q: %w", name, err)
-		}
-	}
-	return dst, nil
-}
-
-// Element represents an element in an object.
-type Element struct {
-	// Name of the element
-	Name string
-	// Type of the element
-	Type Type
-	// Iter containing the element
-	Iter Iter
-}
-
-// Elements contains all elements in an object
-// kept in original order.
-// And index contains lookup for object keys.
-type Elements struct {
-	Elements []Element
-	Index    map[string]int
-}
-
-// Lookup a key in elements and return the element.
-// Returns nil if key doesn't exist.
-// Keys are case sensitive.
-func (e Elements) Lookup(key string) *Element {
-	idx, ok := e.Index[key]
-	if !ok {
-		return nil
-	}
-	return &e.Elements[idx]
-}
-
-// MarshalJSON will marshal the entire remaining scope of the iterator.
-func (e Elements) MarshalJSON() ([]byte, error) {
-	return e.MarshalJSONBuffer(nil)
-}
-
-// MarshalJSONBuffer will marshal all elements.
-// An optional buffer can be provided for fewer allocations.
-// Output will be appended to the destination.
-func (e Elements) MarshalJSONBuffer(dst []byte) ([]byte, error) {
-	dst = append(dst, '{')
-	for i, elem := range e.Elements {
-		dst = append(dst, '"')
-		dst = escapeBytes(dst, []byte(elem.Name))
-		dst = append(dst, '"', ':')
-		var err error
-		dst, err = elem.Iter.MarshalJSONBuffer(dst)
-		if err != nil {
-			return nil, err
-		}
-		if i < len(e.Elements)-1 {
-			dst = append(dst, ',')
-		}
-	}
-	dst = append(dst, '}')
-	return dst, nil
-}
-
-// Parse will return all elements and iterators.
-// An optional destination can be given.
-// The Object will be consumed.
-func (o *Object) Parse(dst *Elements) (*Elements, error) {
-	if dst == nil {
-		dst = &Elements{
-			Elements: make([]Element, 0, 5),
-			Index:    make(map[string]int, 5),
-		}
-	} else {
-		dst.Elements = dst.Elements[:0]
-		for k := range dst.Index {
-			delete(dst.Index, k)
-		}
-	}
-	var tmp Iter
-	for {
-		name, t, err := o.NextElement(&tmp)
-		if err != nil {
-			return nil, err
-		}
-		if t == TypeNone {
-			// Done
-			break
-		}
-		dst.Index[name] = len(dst.Elements)
-		dst.Elements = append(dst.Elements, Element{
-			Name: name,
-			Type: t,
-			Iter: tmp,
-		})
-	}
-	return dst, nil
-}
-
-// NextElement sets dst to the next element and returns the name.
-// TypeNone with nil error will be returned if there are no more elements.
-func (o *Object) NextElement(dst *Iter) (name string, t Type, err error) {
-	if o.off >= len(o.tape.Tape) {
-		return "", TypeNone, nil
-	}
-	// Next must be string or end of object
-	v := o.tape.Tape[o.off]
-	switch Tag(v >> 56) {
-	case TagString:
-		// Read name:
-		name, err = o.tape.stringAt(v)
-		if err != nil {
-			return "", TypeNone, fmt.Errorf("parsing object element name: %w", err)
-		}
-		o.off++
-	case TagObjectEnd:
-		return "", TypeNone, nil
-	default:
-		return "", TypeNone, fmt.Errorf("object: unexpected tag %v", string(v>>56))
-	}
-
-	// Read element type
-	v = o.tape.Tape[o.off]
-	// Move to value (if any)
-	o.off++
-
-	// Set dst
-	dst.cur = v & JSONVALUEMASK
-	dst.t = Tag(v >> 56)
-	dst.off = o.off
-	dst.tape = o.tape
-	dst.calcNext(false)
-	elemSize := dst.addNext
-	dst.calcNext(true)
-	if dst.off+elemSize > len(dst.tape.Tape) {
-		return "", TypeNone, errors.New("element extends beyond tape")
-	}
-	dst.tape.Tape = dst.tape.Tape[:dst.off+elemSize]
-
-	// Skip to next element
-	o.off += elemSize
-	return name, TagToType[dst.t], nil
-}
-
-// Array represents a JSON array.
-// There are methods that allows to get full arrays if the value type is the same.
-// Otherwise an iterator can be retrieved.
-type Array struct {
-	tape ParsedJson
-	off  int
-}
-
-// Iter returns the array as an iterator.
-// This can be used for parsing mixed content arrays.
-// The first value is ready with a call to Next.
-// Calling after last element should have TypeNone.
-func (a *Array) Iter() Iter {
-	i := Iter{
-		tape: a.tape,
-		off:  a.off,
-	}
-	return i
-}
-
-// FirstType will return the type of the first element.
-// If there are no elements, TypeNone is returned.
-func (a *Array) FirstType() Type {
-	iter := a.Iter()
-	return iter.PeekNext()
-}
-
-// MarshalJSON will marshal the entire remaining scope of the iterator.
-func (a *Array) MarshalJSON() ([]byte, error) {
-	return a.MarshalJSONBuffer(nil)
-}
-
-// MarshalJSONBuffer will marshal all elements.
-// An optional buffer can be provided for fewer allocations.
-// Output will be appended to the destination.
-func (a *Array) MarshalJSONBuffer(dst []byte) ([]byte, error) {
-	dst = append(dst, '[')
-	i := a.Iter()
-	var elem Iter
-	for {
-		t, err := i.NextIter(&elem)
-		if err != nil {
-			return nil, err
-		}
-		if t == TypeNone {
-			break
-		}
-		dst, err = elem.MarshalJSONBuffer(dst)
-		if err != nil {
-			return nil, err
-		}
-		if i.PeekNextTag() == TagArrayEnd {
-			break
-		}
-		dst = append(dst, ',')
-	}
-	if i.PeekNextTag() != TagArrayEnd {
-		return nil, errors.New("expected TagArrayEnd as final tag in array")
-	}
-	dst = append(dst, ']')
-	return dst, nil
-}
-
-// Interface returns the array as a slice of interfaces.
-// See Iter.Interface() for a reference on value types.
-func (a *Array) Interface() ([]interface{}, error) {
-	// Estimate length. Assume one value per element.
-	lenEst := (len(a.tape.Tape) - a.off - 1) / 2
-	if lenEst < 0 {
-		lenEst = 0
-	}
-	dst := make([]interface{}, 0, lenEst)
-	i := a.Iter()
-	for i.Next() != TypeNone {
-		elem, err := i.Interface()
-		if err != nil {
-			return nil, err
-		}
-		dst = append(dst, elem)
-	}
-	return dst, nil
-}
-
-// AsFloat returns the array values as float.
-// Integers are automatically converted to float.
-func (a *Array) AsFloat() ([]float64, error) {
-	// Estimate length
-	lenEst := (len(a.tape.Tape) - a.off - 1) / 2
-	if lenEst < 0 {
-		lenEst = 0
-	}
-	dst := make([]float64, 0, lenEst)
-
-readArray:
-	for {
-		tag := Tag(a.tape.Tape[a.off] >> 56)
-		a.off++
-		switch tag {
-		case TagFloat:
-			if len(a.tape.Tape) <= a.off {
-				return nil, errors.New("corrupt input: expected float, but no more values")
-			}
-			dst = append(dst, math.Float64frombits(a.tape.Tape[a.off]))
-		case TagInteger:
-			if len(a.tape.Tape) <= a.off {
-				return nil, errors.New("corrupt input: expected integer, but no more values")
-			}
-			dst = append(dst, float64(int64(a.tape.Tape[a.off])))
-		case TagUint:
-			if len(a.tape.Tape) <= a.off {
-				return nil, errors.New("corrupt input: expected integer, but no more values")
-			}
-			dst = append(dst, float64(a.tape.Tape[a.off]))
-		case TagArrayEnd:
-			break readArray
-		default:
-			return nil, fmt.Errorf("unable to convert type %v to float", tag)
-		}
-		a.off++
-	}
-	return dst, nil
-}
-
-// AsInteger returns the array values as float.
-// Integers are automatically converted to float.
-func (a *Array) AsInteger() ([]int64, error) {
-	// Estimate length
-	lenEst := (len(a.tape.Tape) - a.off - 1) / 2
-	if lenEst < 0 {
-		lenEst = 0
-	}
-	dst := make([]int64, 0, lenEst)
-readArray:
-	for {
-		tag := Tag(a.tape.Tape[a.off] >> 56)
-		a.off++
-		switch tag {
-		case TagFloat:
-			if len(a.tape.Tape) <= a.off {
-				return nil, errors.New("corrupt input: expected float, but no more values")
-			}
-			val := math.Float64frombits(a.tape.Tape[a.off])
-			if val > math.MaxInt64 {
-				return nil, errors.New("float value overflows int64")
-			}
-			if val < math.MinInt64 {
-				return nil, errors.New("float value underflows int64")
-			}
-			dst = append(dst, int64(val))
-		case TagInteger:
-			if len(a.tape.Tape) <= a.off {
-				return nil, errors.New("corrupt input: expected integer, but no more values")
-			}
-			dst = append(dst, int64(a.tape.Tape[a.off]))
-		case TagUint:
-			if len(a.tape.Tape) <= a.off {
-				return nil, errors.New("corrupt input: expected integer, but no more values")
-			}
-
-			val := a.tape.Tape[a.off]
-			if val > math.MaxInt64 {
-				return nil, errors.New("unsigned integer value overflows int64")
-			}
-
-			dst = append(dst)
-		case TagArrayEnd:
-			break readArray
-		default:
-			return nil, fmt.Errorf("unable to convert type %v to integer", tag)
-		}
-		a.off++
-	}
-	return dst, nil
-}
-
-// AsString returns the array values as a slice of strings.
-// No conversion is done.
-func (a *Array) AsString() ([]string, error) {
-	// Estimate length
-	lenEst := len(a.tape.Tape) - a.off - 1
-	if lenEst < 0 {
-		lenEst = 0
-	}
-	dst := make([]string, 0, lenEst)
-	i := a.Iter()
-	var elem Iter
-	for {
-		t, err := i.NextIter(&elem)
-		if err != nil {
-			return nil, err
-		}
-		switch t {
-		case TypeNone:
-			return dst, nil
-		case TypeString:
-			s, err := elem.String()
-			if err != nil {
-
-			}
-			dst = append(dst, s)
-		default:
-			return nil, fmt.Errorf("element in array is not string, but %v", t)
-		}
-	}
 }
 
 func (pj *ParsedJson) Reset() {
